// src/proto/tm2/multisig.ts
export interface PubKeyMultisig {
  k: Long;
  pub_keys: Any[];
}

export interface Multisignature {
  bit_array?: CompactBitArray | undefined;
  sigs: Uint8Array[];
}

export interface CompactBitArray {
  /** The number of extra bits in elems. */
  extra_bits_stored: number;
  elems: Uint8Array;
}

function createBasePubKeyMultisig(): PubKeyMultisig {
  return { k: Long.UZERO, pub_keys: [] };
}

export const PubKeyMultisig: MessageFns<PubKeyMultisig> = {
  encode(message: PubKeyMultisig, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (!message.k.equals(Long.UZERO)) {
      writer.uint32(8).uint64(message.k.toString());
    }
    for (const v of message.pub_keys) {
      Any.encode(v!, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PubKeyMultisig {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePubKeyMultisig();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.k = Long.fromString(reader.uint64().toString(), true);
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.pub_keys.push(Any.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PubKeyMultisig {
    return {
      k: isSet(object.threshold) ? Long.fromValue(object.threshold) : Long.UZERO,
      pub_keys: globalThis.Array.isArray(object?.pubkeys) ? object.pubkeys.map((e: any) => Any.fromJSON(e)) : [],
    };
  },

  toJSON(message: PubKeyMultisig): unknown {
    const obj: any = {};
    if (message.k !== undefined) {
      obj.threshold = (message.k || Long.UZERO).toString();
    }
    if (message.pub_keys?.length) {
      obj.pubkeys = message.pub_keys.map((e) => Any.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PubKeyMultisig>, I>>(base?: I): PubKeyMultisig {
    return PubKeyMultisig.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PubKeyMultisig>, I>>(object: I): PubKeyMultisig {
    const message = createBasePubKeyMultisig();
    message.k = (object.k !== undefined && object.k !== null) ? Long.fromValue(object.k) : Long.UZERO;
    message.pub_keys = object.pub_keys?.map((e) => Any.fromPartial(e)) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends Long ? string | number | Long : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}